Changed behavior on A2 reset.
So far, the A2 RES line was connected to the Pico RUN pin causing it to reset. This reset causes external connections like USB and WiFi to be dropped, which usually is not desirable.

Now, the A2 RES line is connected to GPIO28 and an interrupt handler for the GPIO rising edge is introduced. This allows to freely decide which parts of the Pico are to be reset - if any.