Made PIO get of the bus during reset.
This isn't relevant for a "usual" reset during runtime. But when the Pico is powered via USB while the Apple II is power cycled, then strange things happen on the bus. Shutting down the PIO state machines on entering a reset and completely (re-)initializing them coming out of a reset migitates the problems significantly.