Made firmware hangs impossible.
One of my Apple II machines (a 64kB mainboard ][+ clone) shows extremely high bus activity during shutdown. When the Pico is powered via USB while the Apple II is power cycled, this can lead to the Pico waiting forever for the 'write' PIO SM FIFO. This causes the firmware to hang which needs to be avoided at all costs. The value 32 was chosen as a safe upper limit - it's the amount of Pico cycles the 'write' PIO SM waits for the 6502.