Added RESET level check.
The Apple II power-up Reset circuit can vary between individual machines. And the Pico boot time can vary between indiviual fimrwares, i.e. by the amount of 'time critical' content to be copied from flash to RAM before entering the user code.
Therefore it is possible that the Apple II has already come out of Reset by the time the ARM interrupt handler for the Reset line is installed. In this case the initial rising edge is missed.
Fortunately the solution is simple: Check the Reset level after installing the interrupt handler. If it is already high, then explicitly call the handler. It doesn't hurt if this might lead to calling the handler twice.